+++ ng_trajectory/optimizers/matryoshka/main.py
@@ -8,6 +8,8 @@
 from ng_trajectory.interpolators.utils import *
 import ng_trajectory.plot as ngplot
 
+from ng_trajectory.segmentators.utils import gridCompute
+
 from . import transform
 
 import nevergrad
@@ -21,7 +23,7 @@ from concurrent import futures
 from threading import Lock
 
 # Typing
-from typing import Tuple, Callable, Dict, TextIO
+from typing import Tuple, Callable, Dict, TextIO, List
 
 
 # Global variables
@@ -40,6 +42,7 @@ LOGFILE = None
 VERBOSITY = 3
 FILELOCK = Lock()
 HOLDMAP = None
+GRID = None
 
 
 ######################
@@ -63,6 +66,7 @@ def init(points: numpy.ndarray, group_centers: numpy.ndarray, group_centerline: 
         logging_verbosity: int = 2,
         hold_matryoshka: bool = False,
         plot: bool = False,
+        grid: List[float] = [],
         **kwargs):
     """Initialize variables for Matryoshka transformation.
 
@@ -91,9 +95,10 @@ def init(points: numpy.ndarray, group_centers: numpy.ndarray, group_centerline: 
     logging_verbosity -- index for verbosity of logger, int, default 2
     hold_matryoshka -- whether the Matryoshka should be created only once, bool, default False
     plot -- whether a graphical representation should be created, bool, default False
+    grid -- size of the grid used for the points discretization, 2-float List, computed by default
     **kwargs -- arguments not caught by previous parts
     """
-    global OPTIMIZER, MATRYOSHKA, VALID_POINTS, LOGFILE, VERBOSITY, HOLDMAP
+    global OPTIMIZER, MATRYOSHKA, VALID_POINTS, LOGFILE, VERBOSITY, HOLDMAP, GRID
     global CRITERION, CRITERION_ARGS, INTERPOLATOR, INTERPOLATOR_ARGS, SEGMENTATOR, SEGMENTATOR_ARGS, SELECTOR, SELECTOR_ARGS
 
     # Local to global variables
@@ -133,6 +138,13 @@ def init(points: numpy.ndarray, group_centers: numpy.ndarray, group_centerline: 
         MATRYOSHKA = [ transform.matryoshkaCreate(grouplayers[_i], layers_center[_i], layers_count[_i]) for _i in range(len(_groups)) ]
 
         print ("Matryoshka mapping constructed.")
+
+        if GRID is None:
+            if len(grid) == 2:
+                GRID = grid
+            else:
+                _GRID = gridCompute(points)
+                GRID = [ _GRID, _GRID ]
 
 
     # Optimizer definition
@@ -184,7 +196,7 @@ def _opt(points: numpy.ndarray) -> float:
 
     Note: This function is called after all necessary data is received.
     """
-    global VALID_POINTS, CRITERION, CRITERION_ARGS, INTERPOLATOR, INTERPOLATOR_ARGS, MATRYOSHKA, LOGFILE, FILELOCK, VERBOSITY
+    global VALID_POINTS, CRITERION, CRITERION_ARGS, INTERPOLATOR, INTERPOLATOR_ARGS, MATRYOSHKA, LOGFILE, FILELOCK, VERBOSITY, GRID
 
     # Transform points
     points = [ transform.matryoshkaMap(MATRYOSHKA[i], [p])[0] for i, p in enumerate(points) ]
@@ -198,7 +210,7 @@ def _opt(points: numpy.ndarray) -> float:
     invalid = 0
 
     for _p in _points:
-        if not numpy.any(numpy.all(numpy.abs( numpy.subtract(VALID_POINTS, _p[:2]) ) < [ 0.05, 0.05 ], axis = 1)):
+        if not numpy.any(numpy.all(numpy.abs( numpy.subtract(VALID_POINTS, _p[:2]) ) < GRID, axis = 1)):
             invalid += 1
 
     if ( invalid > 0 ):