+++ ng_trajectory/optimizers/braghin/main.py
@@ -8,6 +8,8 @@
 from ng_trajectory.interpolators.utils import *
 import ng_trajectory.plot as ngplot
 
+from ng_trajectory.segmentators.utils import gridCompute
+
 from . import transform
 
 import nevergrad
@@ -21,7 +23,7 @@ from concurrent import futures
 from threading import Lock
 
 # Typing
-from typing import Tuple, Callable, Dict, TextIO
+from typing import Tuple, Callable, Dict, TextIO, List
 
 
 # Global variables
@@ -40,6 +42,8 @@ LOGFILE = None
 VERBOSITY = 3
 FILELOCK = Lock()
 HOLDMAP = None
+GRID = None
+PENALTY = None
 
 
 ######################
@@ -50,6 +54,7 @@ def init(points: numpy.ndarray, group_centers: numpy.ndarray, group_centerline: 
         budget: int = 100,
         groups: int = 8,
         workers: int = 4,
+        penalty: float = 100,
         criterion: Callable[[numpy.ndarray], float] = lambda x: 0,
         criterion_args: Dict[str, any] = {},
         interpolator: Callable[[numpy.ndarray], numpy.ndarray] = lambda x: x,
@@ -67,6 +72,7 @@ def init(points: numpy.ndarray, group_centers: numpy.ndarray, group_centerline: 
         endpoint_distance: float = 0.2,
         endpoint_accuracy: float = 0.02,
         line_reduction: float = 3,
+        grid: List[float] = [],
         **kwargs):
     """Initialize variables for Braghin's transformation.
 
@@ -77,6 +83,7 @@ def init(points: numpy.ndarray, group_centers: numpy.ndarray, group_centerline: 
     budget -- number of generations of genetic algorithm, int, default 100
     groups -- number of groups to segmentate the track into, int, default 8
     workers -- number of threads for GA, int, default 4
+    penalty -- constant used for increasing the penalty criterion, float, default 100
     criterion -- function to evaluate current criterion, callable (mx2 numpy.ndarray -> float),
                  default 'static 0'
     criterion_args -- arguments for the criterion function, dict, default {}
@@ -99,9 +106,10 @@ def init(points: numpy.ndarray, group_centers: numpy.ndarray, group_centerline: 
     endpoint_distance -- starting distance from the center for transformation, float, default 0.2
     endpoint_accuracy -- accuracy of the center-endpoint distance for transformation, float, default 0.02
     line_reduction -- factor by which the number of line points is lowered before internal interpolation, float, default 3
+    grid -- size of the grid used for the points discretization, 2-float List, computed by default
     **kwargs -- arguments not caught by previous parts
     """
-    global OPTIMIZER, CUTS, VALID_POINTS, LOGFILE, VERBOSITY
+    global OPTIMIZER, CUTS, VALID_POINTS, LOGFILE, VERBOSITY, GRID, PENALTY
     global CRITERION, CRITERION_ARGS, INTERPOLATOR, INTERPOLATOR_ARGS, SEGMENTATOR, SEGMENTATOR_ARGS, SELECTOR, SELECTOR_ARGS
 
     # Local to global variables
@@ -116,6 +124,7 @@ def init(points: numpy.ndarray, group_centers: numpy.ndarray, group_centerline: 
     LOGFILE = logfile
     VERBOSITY = logging_verbosity
     _holdtransform = hold_transform
+    PENALTY = penalty
 
 
     VALID_POINTS = points
@@ -141,6 +150,13 @@ def init(points: numpy.ndarray, group_centers: numpy.ndarray, group_centerline: 
                 ngplot.pointsPlot(numpy.asarray(i))
 
         print ("Braghin's transformation constructed.")
+
+        if GRID is None:
+            if len(grid) == 2:
+                GRID = grid
+            else:
+                _GRID = gridCompute(points)
+                GRID = [ _GRID, _GRID ]
 
 
     # Optimizer definition
@@ -192,7 +208,7 @@ def _opt(points: numpy.ndarray) -> float:
 
     Note: This function is called after all necessary data is received.
     """
-    global VALID_POINTS, CRITERION, CRITERION_ARGS, INTERPOLATOR, INTERPOLATOR_ARGS, CUTS, LOGFILE, FILELOCK, VERBOSITY
+    global VALID_POINTS, CRITERION, CRITERION_ARGS, INTERPOLATOR, INTERPOLATOR_ARGS, CUTS, LOGFILE, FILELOCK, VERBOSITY, GRID, PENALTY
 
     # Transform points
     points = transform.transform(points, CUTS)
@@ -206,7 +222,7 @@ def _opt(points: numpy.ndarray) -> float:
     invalid = 0
 
     for _p in _points:
-        if not numpy.any(numpy.all(numpy.abs( numpy.subtract(VALID_POINTS, _p[:2]) ) < [ 0.05, 0.05 ], axis = 1)):
+        if not numpy.any(numpy.all(numpy.abs( numpy.subtract(VALID_POINTS, _p[:2]) ) < GRID, axis = 1)):
             invalid += 1
 
     if ( invalid > 0 ):
@@ -217,7 +233,7 @@ def _opt(points: numpy.ndarray) -> float:
             if VERBOSITY > 1:
                 print ("invalid:%f" % invalid, file=LOGFILE)
             LOGFILE.flush()
-        return 100 * invalid
+        return PENALTY * invalid
 
     _c = CRITERION(**{**{'points': _points}, **CRITERION_ARGS})
     with FILELOCK: