+++ ng_trajectory/selectors/uniform/main.py
@@ -49,6 +49,15 @@ def trajectoryReduce(points: numpy.ndarray, remain: int) -> numpy.ndarray:
     if ROTATE == 0:
         return points[numpy.linspace(0, len(points)-1, remain, dtype=numpy.int, endpoint=False), :]
 
+    ls = numpy.linspace(0, len(points)-1, remain, dtype=numpy.int, endpoint=False)
+    ls2 = numpy.linspace((ls[1] - ls[0])*ROTATE, len(points)-1+(ls[1] - ls[0])*ROTATE, remain, dtype=numpy.int, endpoint=False)
+
+    # Wrap up the array
+    ls2[ls2>=len(points)] -= len(points)
+
+    return points[ls2, :]
+
+    # Older rotate (absolute)
     # Create linspace
     ls = numpy.linspace(
             (len(points)-1) * ROTATE,