+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+### Changed
+- Selectors: Uniform selector now treats rotate as factor between two consecutive points, instead of rotating the whole path.
 
 ## 1.1.0 - 2021-05-18
 ### Added