+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## Unreleased
 ### Added
 - Segmentators/Utils: `gridCompute()` to obtain square grid size of the points.
+- Substitute not selected algorithms with a stub.
+- Parameter `silent_stub` that suppresses errors when calling stub.
 
 ### Changed
 - Selectors: Uniform selector now treats rotate as factor between two consecutive points, instead of rotating the whole path.